+++ src/jhson.js
@@ -650,7 +650,7 @@
             jsonScope.includeAttributes = function( flag ) {
                 __properties.includeAttributes = getDefaultBoolean( flag, __properties.includeAttributes );
 
-                return this;
+                return jsonScope;
             };
 
             /**
@@ -667,7 +667,7 @@
             jsonScope.includeCssProperties = function( flag ) {
                 __properties.includeCssProperties = getDefaultBoolean( flag, __properties.includeCssProperties );
 
-                return this;
+                return jsonScope;
             };
 
             /**
@@ -684,7 +684,7 @@
             jsonScope.includeText = function( flag ) {
                 __properties.includeText = getDefaultBoolean( flag, __properties.includeText );
 
-                return this;
+                return jsonScope;
             };
 
             /**
@@ -701,7 +701,7 @@
             jsonScope.includeChildren = function( flag ) {
                 __properties.includeChildren = getDefaultBoolean( flag, __properties.includeChildren );
 
-                return this;
+                return jsonScope;
             };
 
             /**
@@ -718,7 +718,7 @@
             jsonScope.friendlyFormat = function( flag ) {
                 __properties.friendlyFormat = getDefaultBoolean( flag, __properties.friendlyFormat );
 
-                return this;
+                return jsonScope;
             };
 
             /**
@@ -735,7 +735,7 @@
             jsonScope.indentSpaces = function( spaces ) {
                 __properties.indentSpaces = getDefaultNumber( spaces, __properties.indentSpaces );
 
-                return this;
+                return jsonScope;
             };
 
             /**
@@ -752,7 +752,7 @@
             jsonScope.ignoreNodeTypes = function( types ) {
                 __properties.ignoreNodeTypes = getDefaultStringOrArray( types, __properties.ignoreNodeTypes );
 
-                return this;
+                return jsonScope;
             };
 
             /**
@@ -769,7 +769,7 @@
             jsonScope.ignoreCssProperties = function( properties ) {
                 __properties.ignoreCssProperties = getDefaultStringOrArray( properties, __properties.ignoreCssProperties );
 
-                return this;
+                return jsonScope;
             };
 
             /**
@@ -786,7 +786,7 @@
             jsonScope.ignoreAttributes = function( attributes ) {
                 __properties.ignoreAttributes = getDefaultStringOrArray( attributes, __properties.ignoreAttributes );
 
-                return this;
+                return jsonScope;
             };
 
             /**
@@ -803,7 +803,7 @@
             jsonScope.generateUniqueMissingIds = function( flag ) {
                 __properties.generateUniqueMissingIds = getDefaultBoolean( flag, __properties.generateUniqueMissingIds );
 
-                return this;
+                return jsonScope;
             };
 
             /**
@@ -890,7 +890,7 @@
             htmlScope.json = function( json ) {
                 __properties.json = getDefaultString( json, __properties.json );
 
-                return this;
+                return htmlScope;
             };
 
             /**
@@ -907,7 +907,7 @@
             htmlScope.templateData = function( templateData ) {
                 __properties.templateData = getDefaultObject( templateData, __properties.templateData );
 
-                return this;
+                return htmlScope;
             };
 
             /**
@@ -924,7 +924,7 @@
             htmlScope.removeOriginalAttributes = function( flag ) {
                 __properties.removeOriginalAttributes = getDefaultBoolean( flag, __properties.removeOriginalAttributes );
 
-                return this;
+                return htmlScope;
             };
 
             /**
@@ -941,7 +941,7 @@
             htmlScope.clearOriginalHTML = function( flag ) {
                 __properties.clearOriginalHTML = getDefaultBoolean( flag, __properties.clearOriginalHTML );
 
-                return this;
+                return htmlScope;
             };
 
             /**
@@ -958,7 +958,7 @@
             htmlScope.addCssToHead = function( flag ) {
                 __properties.addCssToHead = getDefaultBoolean( flag, __properties.addCssToHead );
 
-                return this;
+                return htmlScope;
             };
 
             /**
@@ -975,7 +975,7 @@
             htmlScope.clearCssFromHead = function( flag ) {
                 __properties.clearCssFromHead = getDefaultBoolean( flag, __properties.clearCssFromHead );
 
-                return this;
+                return htmlScope;
             };
 
             /**
@@ -992,7 +992,7 @@
             htmlScope.logTemplateDataWarnings = function( flag ) {
                 __properties.logTemplateDataWarnings = getDefaultBoolean( flag, __properties.logTemplateDataWarnings );
 
-                return this;
+                return htmlScope;
             };
 
             /**
@@ -1009,7 +1009,7 @@
             htmlScope.addAttributes = function( flag ) {
                 __properties.addAttributes = getDefaultBoolean( flag, __properties.addAttributes );
 
-                return this;
+                return htmlScope;
             };
 
             /**
@@ -1026,7 +1026,7 @@
             htmlScope.addCssProperties = function( flag ) {
                 __properties.addCssProperties = getDefaultBoolean( flag, __properties.addCssProperties );
 
-                return this;
+                return htmlScope;
             };
 
             /**
@@ -1043,7 +1043,7 @@
             htmlScope.addText = function( flag ) {
                 __properties.addText = getDefaultBoolean( flag, __properties.addText );
 
-                return this;
+                return htmlScope;
             };
 
             /**
@@ -1060,7 +1060,7 @@
             htmlScope.addChildren = function( flag ) {
                 __properties.addChildren = getDefaultBoolean( flag, __properties.addChildren );
 
-                return this;
+                return htmlScope;
             };
 
             /**